feat(util/printk): new more readable and robust printk implemenation
A from-sratch improved implementation of printk that allows printing
strings larger than the allocated print buffer. It also makes this
buffer statically allocated instead of allocating it on the printing cpu
stack.

Finally, it removes the per-uart implementation uart_puts function as this is
essnetially replicated behavior, only the uart_putc is needed. This
logic is now part of the console itself.

The format printf attribute is also added so that errors in the format
string type identifiers are detected at compile time.

Signed-off-by: Jose Martins <josemartins90@gmail.com>